requirements.txt
streamlit>=1.40.0
plotly>=5.18.0
streamlit-calendar>=1.2.0
python-dotenv>=1.0.0
langchain-google-genai>=2.1.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=3.0.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
pydantic>=2.0.0
typing_extensions>=4.8.0
